requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography
pydantic==2.5.2
pydantic-extra-types==2.4.1
python-dotenv==1.0.0
passlib==1.7.4
bcrypt==4.1.1
email-validator==2.1.0
